+++ source/menu/menu_main.cpp
@@ -70,7 +70,7 @@ void CMenu::_hideMain(bool instant)
 	m_btnMgr.hide(m_mainLblLetter, instant);
 	m_btnMgr.hide(m_mainLblNotice, instant);
 	for(u8 i = 0; i < ARRAY_SIZE(m_mainLblUser); ++i)
-		if(m_mainLblUser[i] != (u16)-1)
+		if(m_mainLblUser[i] != -1)
 			m_btnMgr.hide(m_mainLblUser[i], instant);
 }
 
@@ -136,7 +136,7 @@ void CMenu::_showMain(void)
 	}
 
 	for(u8 i = 0; i < ARRAY_SIZE(m_mainLblUser); ++i)
-		if(m_mainLblUser[i] != (u16)-1)
+		if(m_mainLblUser[i] != -1)
 			m_btnMgr.show(m_mainLblUser[i]);
 
 	if(m_gameList.empty())
